arm: migrate the old kmem.c code to new arch API
mmu.c was a copy of the previous kmem.c code which was arm32 specific.
This commit pares that code down, by implementing the arch_* APIs
outlined in the documentation, and eliminating the old duplicate public
APIs from kmem.c.

To do this, we need to migrate some uint32_t to uintptr_t in the cpu.h,
and we also need to create some stub implementations for functions in
the aarch64 code. We also need to comment out some sections of code
which refer to things we're not yet ready to bring back in.

At this point, both arm and aarch64 can fully boot to the end of main()
in kernel/smain.c.

+++ arch/arm/include/arch/cpu.h
@@ -187,23 +187,23 @@ static inline void ICIALLU(void)
 	set_cpreg2(val, 0, c7, c5, 0);
 }
 
-static inline void set_ttbr0(uint32_t val)
+static inline void set_ttbr0(uintptr_t val)
 {
 	set_cpreg2(val, 0, c2, c0, 0);
 }
-static inline uint32_t get_ttbr0(void)
+static inline uintptr_t get_ttbr0(void)
 {
-	uint32_t val;
+	uintptr_t val;
 	get_cpreg2(val, 0, c2, c0, 0);
 	return val;
 }
-static inline void set_ttbr1(uint32_t val)
+static inline void set_ttbr1(uintptr_t val)
 {
 	set_cpreg2(val, 0, c2, c0, 1);
 }
-static inline uint32_t get_ttbr1(void)
+static inline uintptr_t get_ttbr1(void)
 {
-	uint32_t val;
+	uintptr_t val;
 	get_cpreg2(val, 0, c2, c0, 1);
 	return val;
 }
